udev: hardcode the Wacom EKR into the 27QHDT device group
Most EKR users will use it with the 27QHDT, so let's force it into that group
by default. Note that we force it to the 27QHDT touch device (not the pen
device) because we force the pen device to the touch device too. This way they
all end up in the same group.

There's a range of cases that's not covered (including 27QHD, no
touch) or any case of the EKR being used with another tablet or even the
receiver plugged into another USB port. All these cases need custom udev rules
to assign the right device group, let's just get the default right though.

https://bugs.freedesktop.org/show_bug.cgi?id=102235

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>