touchpad: don't turn on the hysteresis on wobbly touchpads, just log it
Touchpad hardware is terrible enough that trying to figure out when to enable
the hysteresis is a cat-and-mouse game of "hey, why did it enable on my
perfect touchpad" vs "hey, why is my pointer wobbly again". Which is fun but
gets boring after a while, especially when you're on the receiving end of
both.

Let's move the whole thing to the kernel (sort-of, see next paragraph) and
merely log a message when we detect pointer jitter. This at least makes users
aware that something isn't right, they can click on the link, read the
documentation and fix it.

Now, the kernel can't be trusted with the fuzz value (see
1523d8bb2e066bec297f7a03ce4a0d8cada8f383) but we should still set it in the
EV_ABS overrides, that's the only generic place we have for hardware quirks.

So basically what we do now is:
* if the fuzz is 0, that could mean either "no fuzz, I'm perfect" or "shrug,
  aydunno" [1]
* if the fuzz is 0, we enable wobble detection and log a message, telling the
  user to read some documentation. The hysteresis remains off.
* the user overrides the fuzz value of EV_ABS_... in 60-evdev.hwdb (provided
  by systemd [2]) and hopefully provides a PR for that override so others get
  to benefit too
* after reboot, the libinput-model-quirks udev callout detects a nonzero fuzz,
  sets the kernel fuzz to 0 and 'saves' the value to a udev property (see
  1523d8b). libinput sees that property value and uses it as a margin value
  for the hysteresis.

This sounds insane and yes, it is. Welcome to the club, we have comfy jackets.

But wait, how is this different to the LIBINPUT_MODEL_WOBBLY_TOUCHPAD and
LIBINPUT_MODEL_PRECISE_TOUCHPAD approach we had and then ditched in
48473994c8e60189356feae7b7eae25288e5ac28. Well, not that much actually,
although now:
* the kernel doesn't interfere with input events anymore, so there is no
  doubling up of the hysteresis, leading to weird delta changes
* the fuzz is adjustable per-device and per axis, so those devices that need a
  specific value to be less/more responsive can be adjusted as such.
* issues that are caused by dodgier-than-usual hardware, power supply
  issues, sweaty fingers or otherwise different finger conductivity can be
  worked around by a local hwdb entry.
* using the above as failure case fallback, everyone gets to be happy, except
  possibly the kids from [2].

Included is a tool to 'measure' the fuzz required. At this point it just reads
the properties/axes and then does it's best to auto-generate a hwdb entry that
matches the user's hardware and sets a fuzz value on the device. Ideally this
reduces the number of hand-holding required in bugzillas. There are plenty of
things that can go wrong, so our fallback is still to throw up our hands and
point to the documentation.

https://bugs.freedesktop.org/show_bug.cgi?id=98839

[1] bonus points for reading this in a homer simpson voice
[2] or whatever the angry kids use instead of systemd

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>